pyproject.toml
[project]
name = "treesight"
version = "1.0.0"
description = "Geospatial KML-to-satellite imagery pipeline"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.21.0",
    "azure-functions-durable>=1.2.9",
    "azure-storage-blob>=12.23.0",
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
    "pydantic>=2.10.0",
    "fiona>=1.10.0",
    "lxml>=5.3.0",
    "rasterio>=1.4.0",
    "shapely>=2.0.0",
    "pyproj>=3.7.0",
    "pystac-client>=0.8.0",
    "planetary-computer>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["treesight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "scripts"]
markers = [
    "integration: tests requiring a live Azurite instance",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "lxml-stubs>=0.5.1",
]
